* Do not fail when worldtime returns a fraction

+++ src/jsMain/kotlin/at/posselt/pfrpg2e/camping/CampingChat.kt
@@ -52,7 +52,7 @@ fun bindCampingChatEventListeners(game: Game, dispatcher: ActionDispatcher) {
         }
     }
     bindChatClick(".km-pass-time") { _, el ->
-        el.dataset["seconds"]?.toInt()?.let {
+        el.dataset["seconds"]?.toDouble()?.let {
             buildPromise {
                 game.time.advance(it).await()
             }
